==> analyses/01_geoparse-mordecai.ipynb <==
# 01 Geoparsing using mordecai

Analysis reads in title and abstract data from file, runs mordecai and returns a data frame with unique locations for each entry

*Note* This code only works when running on the envname kernel, which was created from the mordecai_env_XXXXXX.yml file (details of modules also in requirements.txt).

## Set up: import modules and read in data

## Set up mordecai geoparser
from mordecai import Geoparser
geo = Geoparser()

# test mordecai
geo.geoparse("I traveled from Oxford to Ottawa.")

## Import other modules
import pandas as pd
import pickle
import os
import sqlite3
from multiprocessing.pool import ThreadPool as Pool

## Read in test geoparsing data
#df = pd.read_csv('C:\\Users\\deviv\\Python-working-folder\\test-mordecai\\data\\raw-data\\test-geocode-text.csv')

# combine title and abstract text into one variable
#df['text'] = df['title'] + df['abstract']


## Read in the relevance predictions
relPredCon = sqlite3.connect("\\".join(["C:","Users","deviv","Python-working-folder","test-mordecai","data","raw-data","sql-databases","relevance-predictions.sqlite"]))

relPred_df = pd.read_sql_query("SELECT * from predRel", relPredCon)

# if needed, check table names:
#cursor = relPredCon.cursor()
#cursor.execute("SELECT name FROM sqlite_master WHERE type='table';")
#print(cursor.fetchall())

# transform analysis_id column
relPred_df.analysis_id = relPred_df.analysis_id.astype(int)

relPredCon.close()
relPred_df.head()

## Read in the metadata for all the unique references
uniqueRefsCon = sqlite3.connect("\\".join(["C:","Users","deviv","Python-working-folder","test-mordecai","data","raw-data","sql-databases","unique-refs_v2.sqlite"]))

## add relevance predictions as a table to the new database
#relPred_df.to_sql("predRel", uniqueRefsCon, if_exists="replace")

# print 
uniqueRefs_df = pd.read_sql_query("SELECT analysis_id, duplicate_id, title, abstract FROM uniquerefs", uniqueRefsCon)

# transform analysis_id column
uniqueRefs_df.analysis_id = uniqueRefs_df.analysis_id.astype(int)

## close connections
uniqueRefsCon.close()

uniqueRefs_df.head()

uniqueRefs_df.shape

## subset relevance predctions to >= 0.5
relPred_df = relPred_df[0.5 <= relPred_df['relevance_mean']]
len(relPred_df)

## JOIN the relevance predictions with the text data
df = relPred_df.set_index('analysis_id').join(uniqueRefs_df.set_index('analysis_id'), how = "left", lsuffix = "_rel", rsuffix = "_refs")
#df = relPred_df.join(uniqueRefs_df, on="analysis_id", how = "left", lsuffix = "_rel", rsuffix = "_refs")
df.head()

df.shape # also get the dimensions -- for some reason I don't have matches for all the relevance predictions 36477 as opposed to 73928??

## Create another column with title and abstract combined to Geoparse
df['text'] = df['title'] + df['abstract']

# replace "the United States" with "United States" to avoid returning the virgin islands
# the (?i) means it is case insensitive
df['text'] = df.text.replace("(?i)the (?i)United (?i)States", "United States", regex=True) 

## alternte, but both ways work
# df['text'] = df['text'].str.replace("the United States", "United States")

## Analysis: Geoparse data and export

## Geoparse the text and add results as a new column "geoparse"
#df["geoparse"] = [geo.geoparse(x) for x in df["text"].astype('str')]

def geoparse_text(text):
    return geo.geoparse(text)

if __name__ = '__main__':
    with Pool() as pool:
        df["geoparse"]=pool.map(geoparse_text,df["text"].astype('str'))


## Format dataframe

# create new formatted dataframe: df_clean
def flat_df(df):
    df_geo = df[df["geoparse"].str.len() != 0] #subset where geoparse string is not empty
    df_geo = df_geo.explode('geoparse') #Transforms each element of a list to a row and replicates index and all other columns. When more than one place name appears it creates more than one row.
    df_geo = pd.concat([df_geo.drop(['geoparse'], axis=1), df_geo['geoparse'].apply(pd.Series)], axis=1) #Extract from dic
    df_geo = pd.concat([df_geo.drop(['geo'], axis=1), df_geo['geo'].apply(pd.Series)], axis=1)
    df_geo = df_geo[df_geo['lat'].notnull()] #Removing empty latitude rows
    df_geo.lat = df_geo.lat.astype(float) #Transforms to float
    df_geo.lon =df_geo.lon.astype(float) #Transforms to float
    return df_geo
    
df_clean = flat_df(df)
print(df_clean.dtypes)
print(df_clean[0:5])
df_clean.head(5)

## Save as a .csv
df_clean.to_csv('C:\\Users\\deviv\\Python-working-folder\\test-mordecai\\outputs\\geoparsed-records.csv', index=False)

## Save as an sqlite database
database = 'C:\\Users\\deviv\\Python-working-folder\\test-mordecai\\outputs\\geoparsed-records.sqlite'
conn = sqlite3.connect(database)
df_clean.to_sql("geoparsed-records", con=conn)
